make the :compiler-args be a projectSetting
workaround SBT ordering inconsistency. Here's a minimal failing case:

  scalaVersion := "2.11.7"

(that's the only line in the build.sbt)

then try to read scalaVersion.value in a buildSettings from an
AutoPlugin. It'll come up as 2.10.5. But it is picked up as 2.11.7 when
called from projectSettings.

Hence, I don't believe the ordering rules in
http://www.scala-sbt.org/0.13/docs/Setting-Initialization.html